requirements.txt
opencv-python
ultralytics
grpcio